Struct reactive_graph_plugin_service_impl::container::PluginContainer

pub struct PluginContainer {
    pub id: Uuid,
    pub stem: String,
    pub path: PathBuf,
    pub state: PluginState,
    pub plugin_declaration: RwLock<Option<PluginDeclaration>>,
    pub proxy: Arc<RwLock<Option<Arc<PluginProxy>>>>,
    pub library: RwLock<Option<Arc<Library>>>,
    pub dependencies: DashSet<PluginDependency>,
}

The plugin container holds the meta information and the library.

Fields
id: Uuid

The id.

stem: String

The filename stem (actually the file_prefix). Does not contain the timestamp and the file extension.

path: PathBuf

The path of the plugin in the file system. The plugin is located in the installation directory.

state: PluginState

The state of the plugin.

plugin_declaration: RwLock<Option<PluginDeclaration>>

The plugin declaration.

proxy: Arc<RwLock<Option<Arc<PluginProxy>>>>

The plugin context.

library: RwLock<Option<Arc<Library>>>

The loaded library.

dependencies: DashSet<PluginDependency>

The dependencies of the plugin.

Implementations

impl PluginContainer

pub fn new(stem: String, path: PathBuf) -> Self

pub fn deploy_dll(&mut self) -> PluginTransitionResult

Moves the file from the deploy directory to the installed directory

pub fn load_dll(&mut self) -> PluginTransitionResult

Loads the dynamic link library into memory.

pub fn unload_dll(&mut self) -> PluginTransitionResult

Unloads the dynamic link library.

pub fn uninstall_dll(&mut self) -> PluginTransitionResult

Uninstalls the dynamic link library by removing the file from the plugin folder.

pub fn load_plugin_declaration(&mut self) -> PluginTransitionResult

Loads the plugin declaration from the dynamic link library. The plugin declaration contains version information and a registration method.

pub fn check_compatibility(&mut self) -> PluginTransitionResult

Checks for compatibility.

pub fn load_plugin_dependencies(&mut self) -> PluginTransitionResult

pub fn construct_proxy( &mut self, plugin_context: Arc<dyn PluginContext + Send + Sync>, ) -> PluginTransitionResult

Constructs the proxy for the plugin.

pub async fn activate(&mut self) -> PluginTransitionResult

Calls the activate method of the plugin.

pub async fn deactivate(&mut self) -> PluginTransitionResult

Calls the deactivate method of the plugin

pub fn remove_proxy(&mut self) -> PluginTransitionResult

Removes the proxy from the plugin.

pub fn start(&mut self) -> Result<(), PluginStartError>

Starts the plugin.

pub fn stop(&mut self) -> Result<(), PluginStopError>

pub fn uninstall(&mut self) -> Result<(), PluginUninstallError>

pub fn redeploy(&mut self) -> Result<(), PluginDeployError>

pub fn disable(&mut self) -> Result<(), PluginDisableError>

pub fn name(&self) -> Option<String>

pub fn name_canonicalized(&self) -> Option<String>

pub fn name_version(&self) -> Option<String>

pub fn description(&self) -> Option<String>

pub fn version(&self) -> Option<String>

pub fn rustc_version(&self) -> Option<String>

pub fn plugin_api_version(&self) -> Option<String>